Add MoviePortrait composable
This composable can be used to display movies in a grid.

An implementation detail is that the model for the Composable wants an ImageRequest to handle the poster. This means the image can to be loaded before the Composable is shown. This makes the "time to first frame" higher, but it means the Composable is fully rendered from the very beginning, as the image is available immediately.