Add PartialDateTime class
Adds a class that represents a date/time value that might not have all the components present.

It's needed to store a watch date of a movie/episode, as the user might not know/remember all details of when something was watched.

Not all combinations of missing data are supported. In ascending order of precision, when a component is unknown, all components after it must be as well. Time is considered a single indivisible component.

Suppoorted types are:
- Local.Year: just a year, e.g. 2024
- Local.YearMonth: year and month, e.g. July 2024
- Local.Date: date, e.g. July 12th, 2024
- Local.DateTime: date and time, e.g. July 12th, 2024 at 21:00:00
- Zoned: composes any Local value with a time zone