Fix DbText.Custom.toCouchTrackerUri() and improve tests
This function was putting the text in a path component rather than the query string, which was wrong.

Tests have been updated to catch this and expanded with property testing.

This also removes the URI from DbText.UknownDefault and restricts having fragments in DbText URIs, so that the URI can always be reconstructed by DbText.UknownDefault.