Add ActionState and use it when saving watched item (#102, #103)
This new state allows to execute arbitrary actions and have a convenient way to keep their state.

The `rememberProfileDbActionState` is a specializion of this for operations on the profile database.

This also introduces a new `Result` type, which is "extended" by `Loadable` and `Actionable`.

For now the loading state simply disables everything, I plan to handle it better for long delays in a follow up PR of #103.