Treat TMDB downloads as Flows
TMDB downloads will now return a FLow. This will allow to respond to changes in the local sqlite cache.

Currently, only the first emitted value is used, but this is a first step to have responsive models.

Since expire/prefetch have effectively the same behavior now (first emit the value in the cache, then try to download), I've removed the concept of prefetching