Move MovieScreenModel to ViewModel
In this commit, the model for the movie screen is refactored to use ViewModel instead. This will preserve the state across screen changes.
Moreover, everything is a flow end-to-end now, so any change in the input language, or a change in the database will be propagated to the UI.

Also, anything that can be loaded independently is now loaded independently. After this PR, no further significant speed improvements are possible.

Note: a similar refactor will be also done for the show screen and for the Watched Item screen.